Create ShortestPalindrome.java
Using KMP Subtring pattern matching , 
This problem can be changed to find the longestPrefix that is a palindrome. then append the reverse of the suffix to the beginning.
using brute force to find the longestPrefix will need O(N^2)
using KMP will be O(N)
using KMP patter=patter+"#"+reverse(pattern)
the lastvalue in the kmp table will be the length of the longest prefix.