Create RemoveDuplicateLetters.java
the visited is necessary since if don't have that, you might have the case that abacb->acb , so it will use the second a to replace the first a
but if a is visted alredy and (evne though a< prev(b)) we don't need to pop the previous a and b, this turns out to be right
since if you have concern look at string "adabdc" so the seocnd a won't get used, but we still return the correct solution,
since when we see b, d get poped we still return abdc. not adbc.